v5.16.9, update imacpro applesmc patch
Fixed the matching order for "iMacPro" to be before iMac and MacPro.

When https://github.com/t2linux/kernel/tree/linux-5.16.y-t2-arch gets
updated, I'll make it so the kernel source comes from there and then it
should only be the patches putting apple-bce and apple-ibridge in tree
that will need to remain here. That might mean there'd end up being a
5.16.9-2 release which would just have some changes to packageing
but should be functionally the same.