compiletest: get version from file
thus you can just edit that file to update the new version.

later we can probably make it create a tag with that version if:

a) no such tag already exists
b) the compiletest doesn't have errors
c) it's on the main branch

we might want to make the version have a -x at the end e.g. 6.1.2-1 so
we can release fixes between versions